refactor(bridge): extract contacts FFI wrapper
